+++ src/main/java/com/parse/gwt/client/Parse.java
@@ -36,6 +36,56 @@ import java.util.Map;
  */
 public class Parse {
 
+    public static class Config {
+        public static void get(final AsyncCallback<com.parse.gwt.client.Config> callback) {
+            Shape.get(Parse.SERVER_URL + "/config")
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            callback.onFailure(throwable);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            try {
+                                com.parse.gwt.client.Config config = com.parse.gwt.client.Config.fromJSON(s);
+                                callback.onSuccess(config);
+                            } catch (Exception e){
+                                callback.onFailure(e);
+                            }
+                        }
+                    });
+        }
+        public static void update(String parameter, JSONValue value, final AsyncCallback<com.parse.gwt.client.Config> callback) {
+            JSONObject params = new JSONObject();
+            params.put(parameter, value);
+            JSONObject body = new JSONObject();
+            body.put("params", params);
+            Shape.put(Parse.SERVER_URL + "/config")
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .body(body.toString())
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            callback.onFailure(throwable);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            try {
+                                com.parse.gwt.client.Config config = com.parse.gwt.client.Config.fromJSON(s);
+                                callback.onSuccess(config);
+                            } catch (Exception e){
+                                callback.onFailure(e);
+                            }
+                        }
+                    });
+        }
+    }
+
     public static class Cloud {
         public static void run(String function, ParseObject object, final AsyncCallback<ParseResponse> callback) {
             Shape.post(Parse.SERVER_URL + "functions/" + function)